refactor: move announcement component into separate file
