+++ app/components/Analyst/Project/Announcements/index.ts
@@ -1,3 +1,4 @@
+export { default as Announcement } from './Announcement';
 export { default as AnnouncementsForm } from './AnnouncementsForm';
 export { default as AnnouncementsHeader } from './AnnouncementsHeader';
 export { default as ViewAnnouncements } from './ViewAnnouncements';
